Add typing, improve style of README, re-order files
- Typehinted most objects and funcs (continue when I'm free)
- Updated README.md
- Added requirements.txt for dependencies
- Updated LICENSE
- Re-organised files and objects/classes
