+++ requirements.txt
@@ -1,0 +1,2 @@
+aiohttp>=3.7.4
+aiolimiter>=1.0.0
