+++ unbelipy/constants.py
@@ -22,38 +22,5 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-class UnbError(Exception):
-    """Base exception class for unbelipy."""
-    pass
-
-class ResponseError(UnbError):
-    pass
-
-class BadRequest(ResponseError):
-    pass
-
-class Unauthorized(ResponseError):
-    pass
-
-class Forbidden(ResponseError):
-    pass
-
-class NotFound(ResponseError):
-    pass
-
-class TooManyRequests(ResponseError):
-    pass
-
-class InternalServerError(ResponseError):
-    pass
-
-class UnknownError(UnbError):
-    pass
-
-API_ERRORS = {
-    400: BadRequest,
-    401: Unauthorized,
-    403: Forbidden,
-    429: TooManyRequests,
-    500: InternalServerError
-}
+API_VERSION = "v1"
+API_BASE_URL = f"https://unbelievaboat.com/api/{API_VERSION}/"
