+++ unbelipy/rate_limits.py
@@ -1,11 +1,41 @@
+"""
+MIT License
+
+Copyright (c) 2021 ChrisDewa
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
 import asyncio
 from collections import defaultdict
-from datetime import datetime, timedelta
-from typing import Dict, List
+from datetime import (
+    datetime, 
+    timedelta
+)
+from typing import (
+    Dict, 
+    List,
+    Any
+)
 
 from aiohttp import ClientResponse
 from aiolimiter import AsyncLimiter
-
 
 class BucketHandler:
     """
@@ -22,27 +52,30 @@ class BucketHandler:
     cond = None
     prevent_429 = False
 
-    def __init__(self, bucket: str):
-        self.bucket = bucket
+    def __init__(self, bucket: str) -> None:
+        self.bucket: str = bucket
 
-    def __repr__(self):
-        return (f"RateLimit(bucket={self.bucket}, limit={self.limit}, remaining={self.remaining}, "
-                f"reset={self.reset}, retry_after={self.retry_after})")
+    def __repr__(self) -> str:
+        return (
+            f"RateLimit(bucket={self.bucket}, limit={self.limit}, remaining={self.remaining}, "
+            f"reset={self.reset}, retry_after={self.retry_after})"
+        )
 
-    def check_limit_headers(self, r: ClientResponse):
-        limits = dict()
-        header_attrs = {
+    def check_limit_headers(self, response: ClientResponse) -> None:
+        limits = {}
+        header_attrs: Dict[str, str] = {
             'X-RateLimit-Limit': 'limit',
             'X-RateLimit-Remaining': 'remaining',
             'X-RateLimit-Reset': 'reset',
         }
         for key in header_attrs:
-            value = r.headers.get(key)
+            value = response.headers.get(key)
             if value is not None:
                 value = int(value)
                 if key == 'X-RateLimit-Reset':
                     value = datetime.utcfromtimestamp(value / 1000)
             limits[header_attrs[key]] = value
+
         for k, v in limits.items():
             setattr(self, k, v)
 
@@ -56,23 +89,21 @@ class BucketHandler:
                 await asyncio.sleep(to_wait)
         return self
 
-    async def __aexit__(self, *args):
+    async def __aexit__(self, *args: Any) -> None:
         if self.prevent_429 is True:
             self.cond.release()
 
-
 class AsyncNonLimiter:
-    async def __aenter__(self):
+    async def __aenter__(self) -> None:
         pass
 
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
+    async def __aexit__(self, exc_type, exc_val, exc_tb) -> None:
         pass
-
 
 class ClientRateLimits:
     buckets: Dict[str, BucketHandler] = dict()
 
-    def __init__(self, prevent_rate_limits: bool):
+    def __init__(self, prevent_rate_limits: bool) -> None:
         self.global_limiter = AsyncLimiter(20, 1) if prevent_rate_limits is True else AsyncNonLimiter()
 
     def currently_limited(self) -> List[str]:
@@ -91,6 +122,5 @@ class ClientRateLimits:
         """
         return any(self.currently_limited())
 
-    def is_limited(self, bucket: str):
+    def is_limited(self, bucket: str) -> bool:
         return bucket in self.currently_limited()
-
